+++ config/livewire-tables.php
@@ -1,8 +1,4 @@
 <?php
 
 return [
-    /**
-     * Options: tailwind | bootstrap-4 | bootstrap-5.
-     */
-    'theme' => 'tailwind',
 ];