+++ src/Traits/WithDateRangeFilter.php
@@ -1,0 +1,10 @@
+<?php
+
+
+namespace Rappasoft\LaravelLivewireTables\Traits;
+
+
+trait WithDateRangeFilter
+{
+
+}