+++ src/Traits/WithFilters.php
@@ -4,9 +4,9 @@ namespace Rappasoft\LaravelLivewireTables\Traits;
 
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Relations\Relation;
+use Rappasoft\LaravelLivewireTables\Columns\Column;
+use Rappasoft\LaravelLivewireTables\Filters\Filter;
 use Rappasoft\LaravelLivewireTables\Utilities\ColumnUtilities;
-use Rappasoft\LaravelLivewireTables\Views\Column;
-use Rappasoft\LaravelLivewireTables\Views\Filter;
 
 /**
  * Trait WithFilters.
@@ -89,7 +89,7 @@ trait WithFilters
     public function checkFilters(): void
     {
         foreach ($this->filters() as $filter => $_default) {
-            if (! isset($this->filters[$filter]) || $this->filters[$filter] === '') {
+            if (!isset($this->filters[$filter]) || $this->filters[$filter] === '') {
                 $this->filters[$filter] = null;
             }
         }
@@ -99,19 +99,14 @@ trait WithFilters
      * Cleans $filter property of any values that don't exist
      * in the filter() definition.
      */
-    public function cleanFilters(): void
+    public function allowFiltersValues(): void
     {
         // Filter $filters values
         $this->filters = collect($this->filters)->filter(function ($filterValue, $filterName) {
             $filterDefinitions = $this->filters();
 
-            // Ignore search
-            if ($filterName === 'search') {
-                return true;
-            }
-
             // Filter out any keys that weren't defined as a filter
-            if (! isset($filterDefinitions[$filterName])) {
+            if (!isset($filterDefinitions[$filterName])) {
                 return false;
             }
 
@@ -120,22 +115,9 @@ trait WithFilters
                 return true;
             }
 
-            // Handle 'select' filters
-            if ($filterDefinitions[$filterName]->isSelect()) {
-                foreach ($this->getFilterOptions($filterName) as $optionValue) {
-                    // If the option is an integer, typecast filter value
-                    if (is_int($optionValue) && $optionValue === (int)$filterValue) {
-                        return true;
-                    }
+            $filter = $filterDefinitions[$filterName];
 
-                    // Strict check the value
-                    if ($optionValue === $filterValue) {
-                        return true;
-                    }
-                }
-            }
-
-            return false;
+            return $filter->allowedValue($filterValue);
         })->toArray();
     }
 
@@ -152,33 +134,37 @@ trait WithFilters
     /**
      * Check if a filter exists and isn't null
      *
-     * @param  string  $filter
+     * @param string $filterName
      *
      * @return bool
      */
-    public function hasFilter(string $filter): bool
+    public function hasFilter(string $filterName): bool
     {
-        return isset($this->filters[$filter]) && $this->filters[$filter] !== null && $this->filters[$filter] !== '';
+        return isset($this->filters[$filterName])
+            && $this->filters[$filterName] !== null
+            && (
+                (is_string($this->filters[$filterName]) && $this->filters[$filterName] !== '')
+                || (is_array($this->filters[$filterName]) && count($this->filters[$filterName]))
+            );
     }
 
     /**
      * Get the value of a given filter
      *
-     * @param  string  $filter
+     * @param string $filterName
      *
      * @return int|string|null
      */
-    public function getFilter(string $filter)
+    public function getFilter(string $filterName)
     {
-        if ($this->hasFilter($filter)) {
-            if (in_array($filter, collect($this->filters())->keys()->toArray(), true) && $this->filters()[$filter]->isSelect()) {
-                return $this->hasIntegerKeys($filter) ? (int)$this->filters[$filter] : trim($this->filters[$filter]);
-            }
-
-            return trim($this->filters[$filter]);
+        if (!$this->hasFilter($filterName)) {
+            return null;
         }
-
-        return null;
+        if (!in_array($filterName, collect($this->filters())->keys()->toArray(), true)) {
+            return null;
+        }
+        $filter = $this->filters()[$filterName];
+        return $filter->processValue($this->filters[$filterName]);
     }
 
     /**
@@ -187,7 +173,7 @@ trait WithFilters
     public function getFilters(): array
     {
         return collect($this->filters)
-            ->reject(fn ($value) => $value === null || $value === '')
+            ->reject(fn($value) => $value === null || $value === '')
             ->toArray();
     }
 
@@ -197,7 +183,7 @@ trait WithFilters
     public function getFiltersWithoutSearch(): array
     {
         return collect($this->getFilters())
-            ->reject(fn ($_value, $key) => $key === 'search')
+            ->reject(fn($_value, $key) => $key === 'search')
             ->toArray();
     }
 
@@ -214,41 +200,13 @@ trait WithFilters
     }
 
     /**
-     * Get the options for a given filter if they exist
-     *
-     * @param  string  $filter
-     *
-     * @return array
-     */
-    public function getFilterOptions(string $filter): array
-    {
-        return collect($this->filters()[$filter]->options())
-            ->keys()
-            ->reject(fn ($item) => $item === '' || $item === null)
-            ->values()
-            ->toArray();
-    }
-
-    /**
-     * Check whether the filter has numeric keys or not
-     *
-     * @param  string  $filter
-     *
-     * @return bool
-     */
-    public function hasIntegerKeys(string $filter): bool
-    {
-        return is_int($this->getFilterOptions($filter)[0] ?? null);
-    }
-
-    /**
      * Collects columns with $searchable = true
      *
      * @return Column[]
      */
-    public function getSearchableColumns() : array
+    public function getSearchableColumns(): array
     {
-        return array_filter($this->columns(), fn (Column $column) => $column->isSearchable());
+        return array_filter($this->columns(), fn(Column $column) => $column->isSearchable());
     }
 
     /**
@@ -277,11 +235,11 @@ trait WithFilters
                     if ($column->hasSearchCallback()) {
                         // Call the callback
                         ($column->getSearchCallback())($subQuery, $search);
-                    } elseif (! $hasRelation || $selectedColumn) { // If the column isn't a relation or if it was previously selected
+                    } elseif (!$hasRelation || $selectedColumn) { // If the column isn't a relation or if it was previously selected
                         $whereColumn = $selectedColumn ?? $column->column();
 
                         // TODO: Skip Aggregates
-                        if (! $hasRelation) {
+                        if (!$hasRelation) {
                             $whereColumn = $query->getModel()->getTable() . '.' . $whereColumn;
                         }